Simplify subscription routers
